llm-rag-service/requirements.txt
flask
langchain
chromadb
python-dotenv
google-generativeai
